move agents back into agent files and tweak some values
